# src/main/resources/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "User Service API",
    "description": "Documentación de la API de Usuarios",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servicio-Usuarios"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/otp": {
      "post": {
        "security": [],
        "tags": ["Autenticación"],
        "summary": "Generar OTP",
        "description": "Genera un código de un solo uso (OTP) y lo envía al correo del usuario.",
        "operationId": "requestOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "$ref": "#/components/responses/UserNotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          },
          "200": {
            "description": "OTP generado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "security": [],
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión",
        "description": "Permite autenticar un usuario con email y contraseña. Devuelve un token JWT si las credenciales son correctas.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "400": {
            "description": "Contraseña incorrecta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "passwordIncorrecta": {
                    "summary": "La contraseña no coincide con la actual",
                    "value": {
                      "status": 400,
                      "message": "La contraseña actual proporcionada es incorrecta",
                      "timestamp": "2025-09-07T16:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/UserNotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "OtpRequest": {
        "type": "object",
        "description": "Solicitud de generación/envío de un código OTP a un correo electrónico",
        "properties": {
          "email": {
            "type": "string",
            "description": "Correo electrónico al cual se enviará el código OTP",
            "example": "usuario@ejemplo.com",
            "maxLength": 50,
            "minLength": 8
          }
        },
        "required": ["email"]
      },
      "OtpResponse": {
        "type": "object",
        "description": "Respuesta con la información del código OTP generado",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Identificador único del OTP en la base de datos",
            "example": 101
          },
          "otp": {
            "type": "integer",
            "format": "int32",
            "description": "Código OTP generado de 6 dígitos",
            "example": 482915,
            "maxLength": 6,
            "minLength": 6
          },
          "user_id": {
            "type": "integer",
            "format": "int32",
            "description": "ID del usuario al que pertenece el OTP",
            "example": 42
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora en que se creó el OTP",
            "example": "2025-09-04T14:32:00.000+00:00"
          },
          "otp_status": {
            "type": "string",
            "description": "Estado actual del OTP (ejemplo: PENDING, USED, EXPIRED)",
            "example": "PENDING"
          },
          "url": {
            "type": "string",
            "description": "url de recuperacion de contraseña",
            "example": "localhost:8080/api/v1/users/1/password"
          }
        },
        "required": ["created_at", "otp", "otp_status", "user_id"]
      },
      "LoginRequest": {
        "type": "object",
        "description": "DTO que representa las credenciales para iniciar sesión en el sistema",
        "properties": {
          "email": {
            "type": "string",
            "description": "Correo electrónico del usuario",
            "example": "usuario@ejemplo.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña de la cuenta del usuario",
            "example": "Passw0rd2025"
          }
        },
        "required": ["email", "password"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["status", "message", "timestamp"]
      },
...
    },
    "responses": {
...
      }, "UserNotFoundError": {
        "description": "Usuario no encontrado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "examples": {
              "usuarioNoEncontrado": {
                "summary": "Usuario con ID inexistente",
                "value": {
                  "status": 404,
                  "message": "Usuario con ID 101 no encontrado",
                  "timestamp": "2025-09-07T14:30:00Z"
                }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Error inesperado en el servidor",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "examples": {
              "errorInterno": {
                "summary": "Error genérico del servidor",
                "value": {
                  "status": 500,
                  "message": "Se produjo un error interno en el servidor",
                  "timestamp": "2025-09-07T15:33:00Z"
                }
              }
            }
          }
        }
      },
      "ServiceUnavailableError": {
        "description": "Error al comunicarse con el servicio de datos",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "examples": {
              "servicioCaido": {
                "summary": "Servicio no disponible",
                "value": {
                  "status": 503,
                  "message": "No se pudo comunicar con el servicio de datos",
                  "timestamp": "2025-09-07T15:34:00Z"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}